Network.AWS.AppStream.CreateFleet

Description

Creates a new fleet.

Creating a Request

createFleet

Arguments

:: Text
cfName

-> Text
cfImageName

-> Text
cfInstanceType

-> ComputeCapacity
cfComputeCapacity

-> CreateFleet

Creates a value of CreateFleet with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cfDomainJoinInfo - The DirectoryName and OrganizationalUnitDistinguishedName values, which are used to join domains for the AppStream 2.0 streaming instances.

cfDisconnectTimeoutInSeconds - The time after disconnection when a session is considered to have ended. If a user who got disconnected reconnects within this timeout interval, the user is connected back to their previous session. The input can be any numeric value in seconds between 60 and 57600.

cfMaxUserDurationInSeconds - The maximum time for which a streaming session can run. The input can be any numeric value in seconds between 600 and 57600.

cfVPCConfig - The VPC configuration for the fleet.

cfDisplayName - The display name of the fleet.

cfEnableDefaultInternetAccess - Enables or disables default internet access for the fleet.

cfDescription - The description of the fleet.

cfName - A unique identifier for the fleet.

cfImageName - Unique name of the image used by the fleet.

cfInstanceType - The instance type of compute resources for the fleet. Fleet instances are launched from this instance type. Available instance types are: * stream.standard.medium * stream.standard.large * stream.compute.large * stream.compute.xlarge * stream.compute.2xlarge * stream.compute.4xlarge * stream.compute.8xlarge * stream.memory.large * stream.memory.xlarge * stream.memory.2xlarge * stream.memory.4xlarge * stream.memory.8xlarge

cfComputeCapacity - The parameters for the capacity allocated to the fleet.

data CreateFleet

Contains the parameters for the new fleet to create.

See: createFleet smart constructor.

Request Lenses

cfDomainJoinInfo :: Lens' CreateFleet (Maybe DomainJoinInfo)

The DirectoryName and OrganizationalUnitDistinguishedName values, which are used to join domains for the AppStream 2.0 streaming instances.

cfDisconnectTimeoutInSeconds :: Lens' CreateFleet (Maybe Int)

The time after disconnection when a session is considered to have ended. If a user who got disconnected reconnects within this timeout interval, the user is connected back to their previous session. The input can be any numeric value in seconds between 60 and 57600.

cfMaxUserDurationInSeconds :: Lens' CreateFleet (Maybe Int)

The maximum time for which a streaming session can run. The input can be any numeric value in seconds between 600 and 57600.

cfVPCConfig :: Lens' CreateFleet (Maybe VPCConfig)

The VPC configuration for the fleet.

cfDisplayName :: Lens' CreateFleet (Maybe Text)

The display name of the fleet.

cfEnableDefaultInternetAccess :: Lens' CreateFleet (Maybe Bool)

Enables or disables default internet access for the fleet.

cfDescription :: Lens' CreateFleet (Maybe Text)

The description of the fleet.

cfName :: Lens' CreateFleet Text

A unique identifier for the fleet.

cfImageName :: Lens' CreateFleet Text

Unique name of the image used by the fleet.

cfInstanceType :: Lens' CreateFleet Text

The instance type of compute resources for the fleet. Fleet instances are launched from this instance type. Available instance types are: * stream.standard.medium * stream.standard.large * stream.compute.large * stream.compute.xlarge * stream.compute.2xlarge * stream.compute.4xlarge * stream.compute.8xlarge * stream.memory.large * stream.memory.xlarge * stream.memory.2xlarge * stream.memory.4xlarge * stream.memory.8xlarge

cfComputeCapacity :: Lens' CreateFleet ComputeCapacity

The parameters for the capacity allocated to the fleet.

Destructuring the Response

createFleetResponse

Arguments

:: Int
cfrsResponseStatus

-> CreateFleetResponse

Creates a value of CreateFleetResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cfrsFleet - The details for the created fleet.

cfrsResponseStatus - -- | The response status code.

data CreateFleetResponse

See: createFleetResponse smart constructor.

Response Lenses

cfrsFleet :: Lens' CreateFleetResponse (Maybe Fleet)

The details for the created fleet.

cfrsResponseStatus :: Lens' CreateFleetResponse Int

- | The response status code.