Network.AWS.AppStream.DeleteStack

Description

Deletes the stack. After this operation completes, the environment can no longer be activated, and any reservations made for the stack are released.

Creating a Request

deleteStack

Arguments

:: Text
dsName

-> DeleteStack

Creates a value of DeleteStack with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dsName - The name of the stack to delete.

data DeleteStack

See: deleteStack smart constructor.

Request Lenses

dsName :: Lens' DeleteStack Text

The name of the stack to delete.

Destructuring the Response

deleteStackResponse

Arguments

:: Int
dsrsResponseStatus

-> DeleteStackResponse

Creates a value of DeleteStackResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dsrsResponseStatus - -- | The response status code.

data DeleteStackResponse

See: deleteStackResponse smart constructor.

Response Lenses

dsrsResponseStatus :: Lens' DeleteStackResponse Int

- | The response status code.