Network.AWS.AppStream.DescribeFleets

Description

If fleet names are provided, this operation describes the specified fleets; otherwise, all the fleets in the account are described.

Creating a Request

describeFleets :: DescribeFleets

Creates a value of DescribeFleets with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dfNextToken - The pagination token to use to retrieve the next page of results for this operation. If this value is null, it retrieves the first page.

dfNames - The fleet names to describe. Use null to describe all the fleets for the AWS account.

data DescribeFleets

See: describeFleets smart constructor.

Request Lenses

dfNextToken :: Lens' DescribeFleets (Maybe Text)

The pagination token to use to retrieve the next page of results for this operation. If this value is null, it retrieves the first page.

dfNames :: Lens' DescribeFleets [Text]

The fleet names to describe. Use null to describe all the fleets for the AWS account.

Destructuring the Response

describeFleetsResponse

Arguments

:: Int
dfsrsResponseStatus

-> DescribeFleetsResponse

Creates a value of DescribeFleetsResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dfsrsNextToken - The pagination token to use to retrieve the next page of results for this operation. If there are no more pages, this value is null.

dfsrsFleets - The list of fleet details.

dfsrsResponseStatus - -- | The response status code.

data DescribeFleetsResponse

See: describeFleetsResponse smart constructor.

Response Lenses

dfsrsNextToken :: Lens' DescribeFleetsResponse (Maybe Text)

The pagination token to use to retrieve the next page of results for this operation. If there are no more pages, this value is null.

dfsrsFleets :: Lens' DescribeFleetsResponse [Fleet]

The list of fleet details.

dfsrsResponseStatus :: Lens' DescribeFleetsResponse Int

- | The response status code.