Network.AWS.AppStream.DescribeStacks

Description

If stack names are not provided, this operation describes the specified stacks; otherwise, all stacks in the account are described. To retrieve the next set of items, pass the nextToken value in a subsequent call to this operation.

Creating a Request

describeStacks :: DescribeStacks

Creates a value of DescribeStacks with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dNextToken - The pagination token to use to retrieve the next page of results for this operation. If this value is null, it retrieves the first page.

dNames - The stack names to describe. Use null to describe all the stacks for the AWS account.

data DescribeStacks

See: describeStacks smart constructor.

Request Lenses

dNextToken :: Lens' DescribeStacks (Maybe Text)

The pagination token to use to retrieve the next page of results for this operation. If this value is null, it retrieves the first page.

dNames :: Lens' DescribeStacks [Text]

The stack names to describe. Use null to describe all the stacks for the AWS account.

Destructuring the Response

describeStacksResponse

Arguments

:: Int
desrsResponseStatus

-> DescribeStacksResponse

Creates a value of DescribeStacksResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

desrsNextToken - The pagination token to use to retrieve the next page of results for this operation. If there are no more pages, this value is null.

desrsStacks - The list of stack details.

desrsResponseStatus - -- | The response status code.

data DescribeStacksResponse

See: describeStacksResponse smart constructor.

Response Lenses

desrsNextToken :: Lens' DescribeStacksResponse (Maybe Text)

The pagination token to use to retrieve the next page of results for this operation. If there are no more pages, this value is null.

desrsStacks :: Lens' DescribeStacksResponse [Stack]

The list of stack details.

desrsResponseStatus :: Lens' DescribeStacksResponse Int

- | The response status code.