Network.AWS.AppStream.DisassociateFleet

Description

Disassociates a fleet from a stack.

Creating a Request

disassociateFleet

Arguments

:: Text
dfFleetName

-> Text
dfStackName

-> DisassociateFleet

Creates a value of DisassociateFleet with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dfFleetName - The name of the fleet to disassociate.

dfStackName - The name of the stack with which the fleet is associated.

data DisassociateFleet

See: disassociateFleet smart constructor.

Request Lenses

dfFleetName :: Lens' DisassociateFleet Text

The name of the fleet to disassociate.

dfStackName :: Lens' DisassociateFleet Text

The name of the stack with which the fleet is associated.

Destructuring the Response

disassociateFleetResponse

Arguments

:: Int
drsResponseStatus

-> DisassociateFleetResponse

Creates a value of DisassociateFleetResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

drsResponseStatus - -- | The response status code.

data DisassociateFleetResponse

See: disassociateFleetResponse smart constructor.

Response Lenses

drsResponseStatus :: Lens' DisassociateFleetResponse Int

- | The response status code.