Network.AWS.AppStream.ListAssociatedStacks

Description

Lists all stacks to which the specified fleet is associated.

Creating a Request

listAssociatedStacks

Arguments

:: Text
lasFleetName

-> ListAssociatedStacks

Creates a value of ListAssociatedStacks with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

lasNextToken - The pagination token to use to retrieve the next page of results for this operation. If this value is null, it retrieves the first page.

lasFleetName - The name of the fleet whose associated stacks are listed.

data ListAssociatedStacks

See: listAssociatedStacks smart constructor.

Request Lenses

lasNextToken :: Lens' ListAssociatedStacks (Maybe Text)

The pagination token to use to retrieve the next page of results for this operation. If this value is null, it retrieves the first page.

lasFleetName :: Lens' ListAssociatedStacks Text

The name of the fleet whose associated stacks are listed.

Destructuring the Response

listAssociatedStacksResponse

Arguments

:: Int
lasrsResponseStatus

-> ListAssociatedStacksResponse

Creates a value of ListAssociatedStacksResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

lasrsNextToken - The pagination token to use to retrieve the next page of results for this operation. If there are no more pages, this value is null.

lasrsNames - The names of associated stacks.

lasrsResponseStatus - -- | The response status code.

data ListAssociatedStacksResponse

The response from a successful operation.

See: listAssociatedStacksResponse smart constructor.

Response Lenses

lasrsNextToken :: Lens' ListAssociatedStacksResponse (Maybe Text)

The pagination token to use to retrieve the next page of results for this operation. If there are no more pages, this value is null.

lasrsNames :: Lens' ListAssociatedStacksResponse [Text]

The names of associated stacks.

lasrsResponseStatus :: Lens' ListAssociatedStacksResponse Int

- | The response status code.