Network.AWS.AppStream.StartFleet

Description

Starts a fleet.

Creating a Request

startFleet

Arguments

:: Text
staName

-> StartFleet

Creates a value of StartFleet with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

staName - The name of the fleet to start.

data StartFleet

See: startFleet smart constructor.

Request Lenses

staName :: Lens' StartFleet Text

The name of the fleet to start.

Destructuring the Response

startFleetResponse

Arguments

:: Int
sfrsResponseStatus

-> StartFleetResponse

Creates a value of StartFleetResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

sfrsResponseStatus - -- | The response status code.

data StartFleetResponse

See: startFleetResponse smart constructor.

Response Lenses

sfrsResponseStatus :: Lens' StartFleetResponse Int

- | The response status code.