Network.AWS.AppStream.Types

Description

Service Configuration

appStream :: Service

API version 2016-12-01 of the Amazon AppStream SDK configuration.

Errors

_InvalidRoleException :: AsError a => Getting (First ServiceError) a ServiceError

The specified role is invalid.

_ResourceAlreadyExistsException :: AsError a => Getting (First ServiceError) a ServiceError

The specified resource already exists.

_IncompatibleImageException :: AsError a => Getting (First ServiceError) a ServiceError

The image does not support storage connectors.

_ConcurrentModificationException :: AsError a => Getting (First ServiceError) a ServiceError

An API error occurred. Wait a few minutes and try again.

_OperationNotPermittedException :: AsError a => Getting (First ServiceError) a ServiceError

The attempted operation is not permitted.

_ResourceNotFoundException :: AsError a => Getting (First ServiceError) a ServiceError

The specified resource was not found.

_InvalidParameterCombinationException :: AsError a => Getting (First ServiceError) a ServiceError

Indicates an incorrect combination of parameters, or a missing parameter.

_ResourceNotAvailableException :: AsError a => Getting (First ServiceError) a ServiceError

The specified resource exists and is not in use, but isn't available.

_LimitExceededException :: AsError a => Getting (First ServiceError) a ServiceError

The requested limit exceeds the permitted limit for an account.

_ResourceInUseException :: AsError a => Getting (First ServiceError) a ServiceError

The specified resource is in use.

AuthenticationType

data AuthenticationType

Constructors

API

Saml

Userpool

FleetAttribute

data FleetAttribute

Fleet attribute.

Constructors

DomainJoinInfo

VPCConfiguration

VPCConfigurationSecurityGroupIds

FleetErrorCode

data FleetErrorCode

Constructors

DomainJoinErrorAccessDenied

DomainJoinErrorDsMachineAccountQuotaExceeded

DomainJoinErrorFileNotFound

DomainJoinErrorInvalidParameter

DomainJoinErrorLogonFailure

DomainJoinErrorMoreData

DomainJoinErrorNoSuchDomain

DomainJoinErrorNotSupported

DomainJoinInternalServiceError

DomainJoinNerrInvalidWorkgroupName

DomainJoinNerrPasswordExpired

DomainJoinNerrWorkstationNotStarted

IAMServiceRoleIsMissing

IAMServiceRoleMissingDescribeSecurityGroupsAction

IAMServiceRoleMissingDescribeSubnetAction

IAMServiceRoleMissingEniCreateAction

IAMServiceRoleMissingEniDeleteAction

IAMServiceRoleMissingEniDescribeAction

ImageNotFound

InternalServiceError

InvalidSubnetConfiguration

NetworkInterfaceLimitExceeded

SecurityGroupsNotFound

SubnetHasInsufficientIPAddresses

SubnetNotFound

FleetState

data FleetState

Constructors

Running

Starting

Stopped

Stopping

ImageState

data ImageState

Constructors

ISAvailable

ISDeleting

ISFailed

ISPending

ImageStateChangeReasonCode

data ImageStateChangeReasonCode

Constructors

ImageBuilderNotAvailable

InternalError

PlatformType

data PlatformType

Constructors

Windows

SessionState

data SessionState

Possible values for the state of a streaming session.

Constructors

Active

Expired

Pending

StackErrorCode

data StackErrorCode

Constructors

SECInternalServiceError

SECStorageConnectorError

StorageConnectorType

data StorageConnectorType

The type of storage connector. The possible values include: HOMEFOLDERS.

Constructors

Homefolders

VisibilityType

data VisibilityType

Constructors

Private

Public

Application

data Application

An entry for a single application in the application catalog.

See: application smart constructor.

application :: Application

Creates a value of Application with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

aEnabled - If there is a problem, an application can be disabled after image creation.

aLaunchPath - The path to the application executable in the instance.

aLaunchParameters - A list of arguments that are passed to the application at launch.

aName - The unique identifier for the application.

aDisplayName - The name of the application shown to the end users.

aMetadata - Additional attributes that describe the application.

aIconURL - The URL for the application icon. This URL may be time-limited.

aEnabled :: Lens' Application (Maybe Bool)

If there is a problem, an application can be disabled after image creation.

aLaunchPath :: Lens' Application (Maybe Text)

The path to the application executable in the instance.

aLaunchParameters :: Lens' Application (Maybe Text)

A list of arguments that are passed to the application at launch.

aName :: Lens' Application (Maybe Text)

The unique identifier for the application.

aDisplayName :: Lens' Application (Maybe Text)

The name of the application shown to the end users.

aMetadata :: Lens' Application (HashMap Text Text)

Additional attributes that describe the application.

aIconURL :: Lens' Application (Maybe Text)

The URL for the application icon. This URL may be time-limited.

ComputeCapacity

data ComputeCapacity

The capacity configuration for the fleet.

See: computeCapacity smart constructor.

computeCapacity

Arguments

:: Int
ccDesiredInstances

-> ComputeCapacity

Creates a value of ComputeCapacity with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ccDesiredInstances - The desired number of streaming instances.

ccDesiredInstances :: Lens' ComputeCapacity Int

The desired number of streaming instances.

ComputeCapacityStatus

data ComputeCapacityStatus

The capacity information for the fleet.

See: computeCapacityStatus smart constructor.

computeCapacityStatus

Arguments

:: Int
ccsDesired

-> ComputeCapacityStatus

Creates a value of ComputeCapacityStatus with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ccsInUse - The number of instances that are being used for streaming.

ccsRunning - The total number of simultaneous streaming instances that are running.

ccsAvailable - The number of currently available instances that can be used to stream sessions.

ccsDesired - The desired number of streaming instances.

ccsInUse :: Lens' ComputeCapacityStatus (Maybe Int)

The number of instances that are being used for streaming.

ccsRunning :: Lens' ComputeCapacityStatus (Maybe Int)

The total number of simultaneous streaming instances that are running.

ccsAvailable :: Lens' ComputeCapacityStatus (Maybe Int)

The number of currently available instances that can be used to stream sessions.

ccsDesired :: Lens' ComputeCapacityStatus Int

The desired number of streaming instances.

DirectoryConfig

data DirectoryConfig

Full directory configuration details, which are used to join domains for the AppStream 2.0 streaming instances.

See: directoryConfig smart constructor.

directoryConfig

Arguments

:: Text
dcDirectoryName

-> DirectoryConfig

Creates a value of DirectoryConfig with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dcCreatedTime - The time stamp when the directory configuration was created within AppStream 2.0.

dcServiceAccountCredentials - The AccountName and AccountPassword of the service account, to be used by the streaming instance to connect to the directory.

dcOrganizationalUnitDistinguishedNames - The list of the distinguished names of organizational units in which to place computer accounts.

dcDirectoryName - The fully qualified name of the directory, such as corp.example.com

dcCreatedTime :: Lens' DirectoryConfig (Maybe UTCTime)

The time stamp when the directory configuration was created within AppStream 2.0.

dcServiceAccountCredentials :: Lens' DirectoryConfig (Maybe ServiceAccountCredentials)

The AccountName and AccountPassword of the service account, to be used by the streaming instance to connect to the directory.

dcOrganizationalUnitDistinguishedNames :: Lens' DirectoryConfig [Text]

The list of the distinguished names of organizational units in which to place computer accounts.

dcDirectoryName :: Lens' DirectoryConfig Text

The fully qualified name of the directory, such as corp.example.com

DomainJoinInfo

data DomainJoinInfo

The DirectoryName and OrganizationalUnitDistinguishedName values, which are used to join domains for the AppStream 2.0 streaming instances.

See: domainJoinInfo smart constructor.

domainJoinInfo :: DomainJoinInfo

Creates a value of DomainJoinInfo with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

djiOrganizationalUnitDistinguishedName - The distinguished name of the organizational unit to place the computer account in.

djiDirectoryName - The fully qualified name of the directory, such as corp.example.com

djiOrganizationalUnitDistinguishedName :: Lens' DomainJoinInfo (Maybe Text)

The distinguished name of the organizational unit to place the computer account in.

djiDirectoryName :: Lens' DomainJoinInfo (Maybe Text)

The fully qualified name of the directory, such as corp.example.com

Fleet

data Fleet

Contains the parameters for a fleet.

See: fleet smart constructor.

fleet

Arguments

:: Text
fARN

-> Text
fName

-> Text
fImageName

-> Text
fInstanceType

-> ComputeCapacityStatus
fComputeCapacityStatus

-> FleetState
fState

-> Fleet

Creates a value of Fleet with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

fDomainJoinInfo - The DirectoryName and OrganizationalUnitDistinguishedName values, which are used to join domains for the AppStream 2.0 streaming instances.

fDisconnectTimeoutInSeconds - The time after disconnection when a session is considered to have ended. If a user who got disconnected reconnects within this timeout interval, the user is connected back to their previous session. The input can be any numeric value in seconds between 60 and 57600.

fMaxUserDurationInSeconds - The maximum time for which a streaming session can run. The value can be any numeric value in seconds between 600 and 57600.

fCreatedTime - The time at which the fleet was created.

fVPCConfig - The VPC configuration for the fleet.

fFleetErrors - The list of fleet errors is appended to this list.

fDisplayName - The name displayed to end users on the AppStream 2.0 portal.

fEnableDefaultInternetAccess - Whether default internet access is enabled for the fleet.

fDescription - The description displayed to end users on the AppStream 2.0 portal.

fARN - The ARN for the fleet.

fName - The name of the fleet.

fImageName - The image used by the fleet.

fInstanceType - The instance type of compute resources for the fleet. The fleet instances are launched from this instance type.

fComputeCapacityStatus - The capacity information for the fleet.

fState - The current state for the fleet.

fDomainJoinInfo :: Lens' Fleet (Maybe DomainJoinInfo)

The DirectoryName and OrganizationalUnitDistinguishedName values, which are used to join domains for the AppStream 2.0 streaming instances.

fDisconnectTimeoutInSeconds :: Lens' Fleet (Maybe Int)

The time after disconnection when a session is considered to have ended. If a user who got disconnected reconnects within this timeout interval, the user is connected back to their previous session. The input can be any numeric value in seconds between 60 and 57600.

fMaxUserDurationInSeconds :: Lens' Fleet (Maybe Int)

The maximum time for which a streaming session can run. The value can be any numeric value in seconds between 600 and 57600.

fCreatedTime :: Lens' Fleet (Maybe UTCTime)

The time at which the fleet was created.

fVPCConfig :: Lens' Fleet (Maybe VPCConfig)

The VPC configuration for the fleet.

fFleetErrors :: Lens' Fleet [FleetError]

The list of fleet errors is appended to this list.

fDisplayName :: Lens' Fleet (Maybe Text)

The name displayed to end users on the AppStream 2.0 portal.

fEnableDefaultInternetAccess :: Lens' Fleet (Maybe Bool)

Whether default internet access is enabled for the fleet.

fDescription :: Lens' Fleet (Maybe Text)

The description displayed to end users on the AppStream 2.0 portal.

fARN :: Lens' Fleet Text

The ARN for the fleet.

fName :: Lens' Fleet Text

The name of the fleet.

fImageName :: Lens' Fleet Text

The image used by the fleet.

fInstanceType :: Lens' Fleet Text

The instance type of compute resources for the fleet. The fleet instances are launched from this instance type.

fComputeCapacityStatus :: Lens' Fleet ComputeCapacityStatus

The capacity information for the fleet.

fState :: Lens' Fleet FleetState

The current state for the fleet.

FleetError

data FleetError

The details of the fleet error.

See: fleetError smart constructor.

fleetError :: FleetError

Creates a value of FleetError with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

feErrorCode - The error code for the fleet error.

feErrorMessage - The error message generated when the fleet has errors.

feErrorCode :: Lens' FleetError (Maybe FleetErrorCode)

The error code for the fleet error.

feErrorMessage :: Lens' FleetError (Maybe Text)

The error message generated when the fleet has errors.

Image

data Image

New streaming instances are booted from images. The image stores the application catalog and is connected to fleets.

See: image smart constructor.

image

Arguments

:: Text
iName

-> Image

Creates a value of Image with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

iState - The image starts in the PENDING state. If image creation succeeds, it moves to AVAILABLE . If image creation fails, it moves to FAILED .

iPlatform - The operating system platform of the image.

iPublicBaseImageReleasedDate - The AWS release date of the public base image. For private images, this date is the release date of the base image from which the image was created.

iStateChangeReason - The reason why the last state change occurred.

iARN - The ARN for the image.

iCreatedTime - The time stamp when the image was created.

iImageBuilderSupported - Whether an image builder can be launched from this image.

iVisibility - The visibility of an image to the user; images can be public or private.

iBaseImageARN - The source image ARN from which this image was created.

iDisplayName - The display name for the image.

iDescription - A meaningful description for the image.

iApplications - The applications associated with an image.

iName - The unique identifier for the image.

iState :: Lens' Image (Maybe ImageState)

The image starts in the PENDING state. If image creation succeeds, it moves to AVAILABLE . If image creation fails, it moves to FAILED .

iPlatform :: Lens' Image (Maybe PlatformType)

The operating system platform of the image.

iPublicBaseImageReleasedDate :: Lens' Image (Maybe UTCTime)

The AWS release date of the public base image. For private images, this date is the release date of the base image from which the image was created.

iStateChangeReason :: Lens' Image (Maybe ImageStateChangeReason)

The reason why the last state change occurred.

iARN :: Lens' Image (Maybe Text)

The ARN for the image.

iCreatedTime :: Lens' Image (Maybe UTCTime)

The time stamp when the image was created.

iImageBuilderSupported :: Lens' Image (Maybe Bool)

Whether an image builder can be launched from this image.

iVisibility :: Lens' Image (Maybe VisibilityType)

The visibility of an image to the user; images can be public or private.

iBaseImageARN :: Lens' Image (Maybe Text)

The source image ARN from which this image was created.

iDisplayName :: Lens' Image (Maybe Text)

The display name for the image.

iDescription :: Lens' Image (Maybe Text)

A meaningful description for the image.

iApplications :: Lens' Image [Application]

The applications associated with an image.

iName :: Lens' Image Text

The unique identifier for the image.

ImageStateChangeReason

data ImageStateChangeReason

The reason why the last state change occurred.

See: imageStateChangeReason smart constructor.

imageStateChangeReason :: ImageStateChangeReason

Creates a value of ImageStateChangeReason with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

iscrCode - The state change reason code of the image.

iscrMessage - The state change reason message to the end user.

iscrCode :: Lens' ImageStateChangeReason (Maybe ImageStateChangeReasonCode)

The state change reason code of the image.

iscrMessage :: Lens' ImageStateChangeReason (Maybe Text)

The state change reason message to the end user.

ServiceAccountCredentials

data ServiceAccountCredentials

The AccountName and AccountPassword of the service account, to be used by the streaming instance to connect to the directory.

See: serviceAccountCredentials smart constructor.

serviceAccountCredentials

Arguments

:: Text
sacAccountName

-> Text
sacAccountPassword

-> ServiceAccountCredentials

Creates a value of ServiceAccountCredentials with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

sacAccountName - The user name of an account in the directory that is used by AppStream 2.0 streaming instances to connect to the directory. This account must have the following privileges: create computer objects, join computers to the domain, change/reset the password on descendant computer objects for the organizational units specified.

sacAccountPassword - The password for the user account for directory actions.

sacAccountName :: Lens' ServiceAccountCredentials Text

The user name of an account in the directory that is used by AppStream 2.0 streaming instances to connect to the directory. This account must have the following privileges: create computer objects, join computers to the domain, change/reset the password on descendant computer objects for the organizational units specified.

sacAccountPassword :: Lens' ServiceAccountCredentials Text

The password for the user account for directory actions.

Session

data Session

Contains the parameters for a streaming session.

See: session smart constructor.

session

Arguments

:: Text
sId

-> Text
sUserId

-> Text
sStackName

-> Text
sFleetName

-> SessionState
sState

-> Session

Creates a value of Session with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

sAuthenticationType - The authentication method of the user for whom the session was created. It can be API for a user authenticated using a streaming URL or SAML for a SAML federated user.

sId - The unique ID for a streaming session.

sUserId - The identifier of the user for whom the session was created.

sStackName - The name of the stack for which the streaming session was created.

sFleetName - The name of the fleet for which the streaming session was created.

sState - The current state of the streaming session.

sAuthenticationType :: Lens' Session (Maybe AuthenticationType)

The authentication method of the user for whom the session was created. It can be API for a user authenticated using a streaming URL or SAML for a SAML federated user.

sId :: Lens' Session Text

The unique ID for a streaming session.

sUserId :: Lens' Session Text

The identifier of the user for whom the session was created.

sStackName :: Lens' Session Text

The name of the stack for which the streaming session was created.

sFleetName :: Lens' Session Text

The name of the fleet for which the streaming session was created.

sState :: Lens' Session SessionState

The current state of the streaming session.

Stack

data Stack

Details about a stack.

See: stack smart constructor.

stack

Arguments

:: Text
sName

-> Stack

Creates a value of Stack with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

sARN - The ARN of the stack.

sCreatedTime - The time stamp when the stack was created.

sStorageConnectors - The storage connectors to be enabled for the stack.

sDisplayName - A display name for the stack.

sStackErrors - The list of errors associated with the stack.

sDescription - A meaningful description for the stack.

sName - The unique identifier of the stack.

sARN :: Lens' Stack (Maybe Text)

The ARN of the stack.

sCreatedTime :: Lens' Stack (Maybe UTCTime)

The time stamp when the stack was created.

sStorageConnectors :: Lens' Stack [StorageConnector]

The storage connectors to be enabled for the stack.

sDisplayName :: Lens' Stack (Maybe Text)

A display name for the stack.

sStackErrors :: Lens' Stack [StackError]

The list of errors associated with the stack.

sDescription :: Lens' Stack (Maybe Text)

A meaningful description for the stack.

sName :: Lens' Stack Text

The unique identifier of the stack.

StackError

data StackError

Contains the parameters for a stack error.

See: stackError smart constructor.

stackError :: StackError

Creates a value of StackError with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

seErrorCode - The error code of a stack error.

seErrorMessage - The error message of a stack error.

seErrorCode :: Lens' StackError (Maybe StackErrorCode)

The error code of a stack error.

seErrorMessage :: Lens' StackError (Maybe Text)

The error message of a stack error.

StorageConnector

data StorageConnector

Contains the parameters for a storage connector.

See: storageConnector smart constructor.

storageConnector

Arguments

:: StorageConnectorType
scConnectorType

-> StorageConnector

Creates a value of StorageConnector with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

scResourceIdentifier - The ARN associated with the storage connector.

scConnectorType - The type of storage connector. The possible values include: HOMEFOLDERS.

scResourceIdentifier :: Lens' StorageConnector (Maybe Text)

The ARN associated with the storage connector.

scConnectorType :: Lens' StorageConnector StorageConnectorType

The type of storage connector. The possible values include: HOMEFOLDERS.

VPCConfig

data VPCConfig

VPC configuration information.

See: vpcConfig smart constructor.

vpcConfig :: VPCConfig

Creates a value of VPCConfig with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

vcSecurityGroupIds - Security groups associated with the fleet.

vcSubnetIds - The list of subnets to which a network interface is established from the fleet instance.

vcSecurityGroupIds :: Lens' VPCConfig [Text]

Security groups associated with the fleet.

vcSubnetIds :: Lens' VPCConfig [Text]

The list of subnets to which a network interface is established from the fleet instance.