Network.AWS.AppStream.UpdateFleet

Description

Updates an existing fleet. All the attributes except the fleet name can be updated in the STOPPED state. When a fleet is in the RUNNING state, only DisplayName and ComputeCapacity can be updated. A fleet cannot be updated in a status of STARTING or STOPPING .

Creating a Request

updateFleet

Arguments

:: Text
ufName

-> UpdateFleet

Creates a value of UpdateFleet with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ufDomainJoinInfo - The DirectoryName and OrganizationalUnitDistinguishedName values, which are used to join domains for the AppStream 2.0 streaming instances.

ufDisconnectTimeoutInSeconds - The time after disconnection when a session is considered to have ended. If a user who got disconnected reconnects within this timeout interval, the user is connected back to their previous session. The input can be any numeric value in seconds between 60 and 57600.

ufMaxUserDurationInSeconds - The maximum time for which a streaming session can run. The input can be any numeric value in seconds between 600 and 57600.

ufAttributesToDelete - Fleet attributes to be deleted.

ufDeleteVPCConfig - Delete the VPC association for the specified fleet.

ufInstanceType - The instance type of compute resources for the fleet. Fleet instances are launched from this instance type. Available instance types are: * stream.standard.medium * stream.standard.large * stream.compute.large * stream.compute.xlarge * stream.compute.2xlarge * stream.compute.4xlarge * stream.compute.8xlarge * stream.memory.large * stream.memory.xlarge * stream.memory.2xlarge * stream.memory.4xlarge * stream.memory.8xlarge

ufVPCConfig - The VPC configuration for the fleet.

ufDisplayName - The name displayed to end users on the AppStream 2.0 portal.

ufEnableDefaultInternetAccess - Enables or disables default internet access for the fleet.

ufImageName - The image name from which a fleet is created.

ufDescription - The description displayed to end users on the AppStream 2.0 portal.

ufComputeCapacity - The parameters for the capacity allocated to the fleet.

ufName - The name of the fleet.

data UpdateFleet

See: updateFleet smart constructor.

Request Lenses

ufDomainJoinInfo :: Lens' UpdateFleet (Maybe DomainJoinInfo)

The DirectoryName and OrganizationalUnitDistinguishedName values, which are used to join domains for the AppStream 2.0 streaming instances.

ufDisconnectTimeoutInSeconds :: Lens' UpdateFleet (Maybe Int)

The time after disconnection when a session is considered to have ended. If a user who got disconnected reconnects within this timeout interval, the user is connected back to their previous session. The input can be any numeric value in seconds between 60 and 57600.

ufMaxUserDurationInSeconds :: Lens' UpdateFleet (Maybe Int)

The maximum time for which a streaming session can run. The input can be any numeric value in seconds between 600 and 57600.

ufAttributesToDelete :: Lens' UpdateFleet [FleetAttribute]

Fleet attributes to be deleted.

ufDeleteVPCConfig :: Lens' UpdateFleet (Maybe Bool)

Delete the VPC association for the specified fleet.

ufInstanceType :: Lens' UpdateFleet (Maybe Text)

The instance type of compute resources for the fleet. Fleet instances are launched from this instance type. Available instance types are: * stream.standard.medium * stream.standard.large * stream.compute.large * stream.compute.xlarge * stream.compute.2xlarge * stream.compute.4xlarge * stream.compute.8xlarge * stream.memory.large * stream.memory.xlarge * stream.memory.2xlarge * stream.memory.4xlarge * stream.memory.8xlarge

ufVPCConfig :: Lens' UpdateFleet (Maybe VPCConfig)

The VPC configuration for the fleet.

ufDisplayName :: Lens' UpdateFleet (Maybe Text)

The name displayed to end users on the AppStream 2.0 portal.

ufEnableDefaultInternetAccess :: Lens' UpdateFleet (Maybe Bool)

Enables or disables default internet access for the fleet.

ufImageName :: Lens' UpdateFleet (Maybe Text)

The image name from which a fleet is created.

ufDescription :: Lens' UpdateFleet (Maybe Text)

The description displayed to end users on the AppStream 2.0 portal.

ufComputeCapacity :: Lens' UpdateFleet (Maybe ComputeCapacity)

The parameters for the capacity allocated to the fleet.

ufName :: Lens' UpdateFleet Text

The name of the fleet.

Destructuring the Response

updateFleetResponse

Arguments

:: Int
ufrsResponseStatus

-> UpdateFleetResponse

Creates a value of UpdateFleetResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ufrsFleet - A list of fleet details.

ufrsResponseStatus - -- | The response status code.

data UpdateFleetResponse

See: updateFleetResponse smart constructor.

Response Lenses

ufrsFleet :: Lens' UpdateFleetResponse (Maybe Fleet)

A list of fleet details.

ufrsResponseStatus :: Lens' UpdateFleetResponse Int

- | The response status code.